Move example into s3 mod; Move config into store; split fs and s3 config
